Split the source into several files.
This is a pure refactoring to split the source code into several smaller files rather than one large file.
